+++ src/main/java/uk/ac/hope/mcse/segh/servermanagement/repo/HardwareReadingRepository.java
@@ -11,5 +11,5 @@ public interface HardwareReadingRepository extends MongoRepository <HardwareRead
 
     Optional<HardwareReading> findHardwareReadingBy_id(BigInteger _id);
 
-    Optional<HardwareReading> findHardwareReadingByCpu(int cpu);
+    Optional<HardwareReading> findHardwareReadingByCpu(double cpu);
 }
